refactored and included rust

+++ hackerrank/30 days of code/day 25, running time and complexity/running_time_and_complexity.cpp
@@ -1,18 +1,22 @@
+// https://www.hackerrank.com/challenges/30-running-time-and-complexity/problem?isFullScreen=true
+
 #include <iostream>
 #include <cmath>
+
 using namespace std;
 
-int main() {
+int main()
+{
     int n;
     cin >> n;
 
-    for (int number; n-- && cin >> number; ) {
+    for (int number; n-- && cin >> number;) {
         if (number == 1) {
             cout << "Not prime\n";
             continue;
         }
 
-        int i {2}, limit = static_cast<int>(sqrt(number));
+        int i {2}, limit = sqrt(number);
         for (; i <= limit; ++i)
             if (!(number % i)) {
                 cout << "Not prime\n";
@@ -22,5 +26,6 @@ int main() {
         if (i > limit)
             cout << "Prime\n";
     }
+
     return 0;
 }
